Some fixes for 18Rhl
1. RhE corrections
 * When setting par after buying private 6, the par can only be
   70, 75 or 80.
 * Money for presidency will be the winning bid of private 6.
 * Private 6 text updated.
2. Remove (incorrectly located) fixture from pre-alpha testing
 of 18Rhl as changes breaks this fixture. As it is already
 paring of RhE after private auction it is not possible to fix.
 A new fixture will be added (but maybe not until after alpha
 started?).
3. Split out entities (and some related methods) to an entities file.
4. Split out map and tiles (and some related methods) to a map file.
5. Correct special track for privates 2-4 (and 1 if using variant)
 so that these can be use to lay a track between regular track build
 and train buy.

Also removed two unused methods from 18MS.
