+++ lib/engine/game/g_18_rhl/step/buy_train.rb
@@ -11,6 +11,16 @@ module Engine
             # Corps in receivership cannot buy/sell trains (Rule 13.2)
             super.reject { |t| t.owner != @game.depot && (t.owner.receivership? || entity.receivership?) }
           end
+
+          def process_buy_train(action)
+            @round.bought_trains << action.entity
+            super
+          end
+
+          # Need to keep track of bought trains to avoid special tile lay after train buy
+          def round_state
+            { bought_trains: [] }
+          end
         end
       end
     end
